+++ extensions/library-genesis/src/types.ts
@@ -1,16 +1,25 @@
 export interface BookEntry {
   title: string;
   author: string;
-  year: string;
+  year?: string;
   url: string;
-  pages: string;
-  language: string;
-  publisher: string;
-  commentary: string;
   coverUrl: string;
-  extension: string;
-  fileSize: string;
-  md5: string;
-  timeAdded: string;
-  timeLastModified: string;
+  pages?: string;
+  language?: string;
+  publisher?: string;
+  edition?: string;
+  commentary?: string;
+  extension?: string;
+  fileSize?: string;
+  md5?: string;
+  timeAdded?: string;
+  timeLastModified?: string;
+  id?: string;
+  isbn?: string;
+  series?: string;
+  periodical?: string;
+}
+
+export interface LibgenSearchParams {
+  req: string;
 }