+++ extensions/library-genesis/CHANGELOG.md
@@ -1,5 +1,11 @@
 # Library Genesis Changelog
 
+## [Updated search books] - 2022-07-16
+
+- Deprecated [libgen.js](https://github.com/dunn/libgen.js/) due to poor search results.
+- Migrated to a custom scraper that parse raw html from a query request to libgen.
+- Changed the default url to open in browser to the download page for convenience.
+
 ## [Added Library Genesis] - 2022-07-14
 
 Initial version code